Part of a 'manual' 'rebase with history.'

+++ Library/Homebrew/cleanup.rb
@@ -1,177 +1,178 @@
+# typed: false
+# frozen_string_literal: true
+
 require "utils/bottles"
+
 require "formula"
 require "cask/cask_loader"
 require "set"
 
-CLEANUP_DEFAULT_DAYS = 30
-CLEANUP_MAX_AGE_DAYS = 120
+module Homebrew
+  # Helper class for cleaning up the Homebrew cache.
+  #
+  # @api private
+  class Cleanup
+    CLEANUP_DEFAULT_DAYS = Homebrew::EnvConfig.cleanup_periodic_full_days.to_i.freeze
+    private_constant :CLEANUP_DEFAULT_DAYS
 
-module CleanupRefinement
-  refine Enumerator do
-    def parallel
-      queue = Queue.new
+    # {Pathname} refinement with helper functions for cleaning up files.
+    module CleanupRefinement
+      refine Pathname do
+        def incomplete?
+          extname.end_with?(".incomplete")
+        end
 
-      each do |element|
-        queue.enq(element)
-      end
+        def nested_cache?
+          directory? && %w[
+            cargo_cache
+            go_cache
+            go_mod_cache
+            glide_home
+            java_cache
+            npm_cache
+            gclient_cache
+          ].include?(basename.to_s)
+        end
 
-      workers = (0...Hardware::CPU.cores).map do
-        Thread.new do
-          Kernel.loop do
-            begin
-              yield queue.deq(true)
-            rescue ThreadError
-              break # if queue is empty
-            end
+        def go_cache_directory?
+          # Go makes its cache contents read-only to ensure cache integrity,
+          # which makes sense but is something we need to undo for cleanup.
+          directory? && %w[go_cache go_mod_cache].include?(basename.to_s)
+        end
+
+        def prune?(days)
+          return false unless days
+          return true if days.zero?
+
+          return true if symlink? && !exist?
+
+          mtime < days.days.ago && ctime < days.days.ago
+        end
+
+        def stale?(scrub: false)
+          return false unless resolved_path.file?
+
+          if dirname.basename.to_s == "Cask"
+            stale_cask?(scrub)
+          else
+            stale_formula?(scrub)
           end
         end
-      end
 
-      workers.each(&:join)
-    end
-  end
+        private
 
-  refine Pathname do
-    def incomplete?
-      extname.end_with?(".incomplete")
-    end
+        def stale_formula?(scrub)
+          return false unless HOMEBREW_CELLAR.directory?
 
-    def nested_cache?
-      directory? && %w[cargo_cache go_cache glide_home java_cache npm_cache gclient_cache].include?(basename.to_s)
-    end
+          version = if HOMEBREW_BOTTLES_EXTNAME_REGEX.match?(to_s)
+            begin
+              Utils::Bottles.resolve_version(self)
+            rescue
+              nil
+            end
+          end
 
-    def go_cache_directory?
-      # Go makes its cache contents read-only to ensure cache integrity,
-      # which makes sense but is something we need to undo for cleanup.
-      directory? && %w[go_cache].include?(basename.to_s)
-    end
+          version ||= basename.to_s[/\A.*(?:--.*?)*--(.*?)#{Regexp.escape(extname)}\Z/, 1]
+          version ||= basename.to_s[/\A.*--?(.*?)#{Regexp.escape(extname)}\Z/, 1]
 
-    def prune?(days)
-      return false unless days
-      return true if days.zero?
+          return false unless version
 
-      return true if symlink? && !exist?
+          version = Version.new(version)
 
-      mtime < days.days.ago && ctime < days.days.ago
-    end
+          return false unless (formula_name = basename.to_s[/\A(.*?)(?:--.*?)*--?(?:#{Regexp.escape(version)})/, 1])
 
-    def stale?(scrub = false)
-      return false unless resolved_path.file?
+          formula = begin
+            Formulary.from_rack(HOMEBREW_CELLAR/formula_name)
+          rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
+            nil
+          end
 
-      stale_formula?(scrub) || stale_cask?(scrub)
-    end
+          return false if formula.blank?
 
-    private
+          resource_name = basename.to_s[/\A.*?--(.*?)--?(?:#{Regexp.escape(version)})/, 1]
 
-    def stale_formula?(scrub)
-      return false unless HOMEBREW_CELLAR.directory?
+          if resource_name == "patch"
+            patch_hashes = formula.stable&.patches&.select(&:external?)&.map(&:resource)&.map(&:version)
+            return true unless patch_hashes&.include?(Checksum.new(version.to_s))
+          elsif resource_name && (resource_version = formula.stable&.resources&.dig(resource_name)&.version)
+            return true if resource_version != version
+          elsif version.is_a?(PkgVersion)
+            return true if formula.pkg_version > version
+          elsif formula.version > version
+            return true
+          end
 
-      version = if to_s.match?(Pathname::BOTTLE_EXTNAME_RX)
-        begin
-          Utils::Bottles.resolve_version(self)
-        rescue
-          nil
+          return true if scrub && !formula.latest_version_installed?
+
+          return true if Utils::Bottles.file_outdated?(formula, self)
+
+          false
+        end
+
+        def stale_cask?(scrub)
+          return false unless (name = basename.to_s[/\A(.*?)--/, 1])
+
+          cask = begin
+            Cask::CaskLoader.load(name)
+          rescue Cask::CaskError
+            nil
+          end
+
+          return false if cask.blank?
+
+          return true unless basename.to_s.match?(/\A#{Regexp.escape(name)}--#{Regexp.escape(cask.version)}\b/)
+
+          return true if scrub && cask.versions.exclude?(cask.version)
+
+          if cask.version.latest?
+            return mtime < CLEANUP_DEFAULT_DAYS.days.ago &&
+                   ctime < CLEANUP_DEFAULT_DAYS.days.ago
+          end
+
+          false
         end
       end
-
-      version ||= basename.to_s[/\A.*(?:\-\-.*?)*\-\-(.*?)#{Regexp.escape(extname)}\Z/, 1]
-      version ||= basename.to_s[/\A.*\-\-?(.*?)#{Regexp.escape(extname)}\Z/, 1]
-
-      return false unless version
-
-      version = Version.new(version)
-
-      return false unless formula_name = basename.to_s[/\A(.*?)(?:\-\-.*?)*\-\-?(?:#{Regexp.escape(version)})/, 1]
-
-      formula = begin
-        Formulary.from_rack(HOMEBREW_CELLAR/formula_name)
-      rescue FormulaUnavailableError, TapFormulaAmbiguityError, TapFormulaWithOldnameAmbiguityError
-        return false
-      end
-
-      resource_name = basename.to_s[/\A.*?\-\-(.*?)\-\-?(?:#{Regexp.escape(version)})/, 1]
-
-      if resource_name == "patch"
-        patch_hashes = formula.stable&.patches&.select(&:external?)&.map(&:resource)&.map(&:version)
-        return true unless patch_hashes&.include?(Checksum.new(:sha256, version.to_s))
-      elsif resource_name && resource_version = formula.stable&.resources&.dig(resource_name)&.version
-        return true if resource_version != version
-      elsif version.is_a?(PkgVersion)
-        return true if formula.pkg_version > version
-      elsif formula.version > version
-        return true
-      end
-
-      return true if scrub && !formula.installed?
-
-      return true if Utils::Bottles.file_outdated?(formula, self)
-
-      false
     end
 
-    def stale_cask?(scrub)
-      return false unless name = basename.to_s[/\A(.*?)\-\-/, 1]
+    using CleanupRefinement
 
-      cask = begin
-        Cask::CaskLoader.load(name)
-      rescue Cask::CaskUnavailableError
-        return false
-      end
-
-      unless basename.to_s.match?(/\A#{Regexp.escape(name)}\-\-#{Regexp.escape(cask.version)}\b/)
-        return true
-      end
-
-      return true if scrub && !cask.versions.include?(cask.version)
-
-      if cask.version.latest?
-        return mtime < CLEANUP_DEFAULT_DAYS.days.ago &&
-               ctime < CLEANUP_DEFAULT_DAYS.days.ago
-      end
-
-      false
-    end
-  end
-end
-
-using CleanupRefinement
-
-module Homebrew
-  class Cleanup
     extend Predicable
 
-    PERIODIC_CLEAN_FILE = HOMEBREW_CACHE/".cleaned"
+    PERIODIC_CLEAN_FILE = (HOMEBREW_CACHE/".cleaned").freeze
 
-    attr_predicate :dry_run?, :scrub?
-    attr_reader :args, :days, :cache
-    attr_reader :disk_cleanup_size
+    attr_predicate :dry_run?, :scrub?, :prune?
+    attr_reader :args, :days, :cache, :disk_cleanup_size
 
     def initialize(*args, dry_run: false, scrub: false, days: nil, cache: HOMEBREW_CACHE)
       @disk_cleanup_size = 0
       @args = args
       @dry_run = dry_run
       @scrub = scrub
-      @days = days || CLEANUP_MAX_AGE_DAYS
+      @prune = days.present?
+      @days = days || Homebrew::EnvConfig.cleanup_max_age_days.to_i
       @cache = cache
       @cleaned_up_paths = Set.new
     end
 
     def self.install_formula_clean!(f)
-      return if ENV["HOMEBREW_NO_INSTALL_CLEANUP"]
-      return unless ENV["HOMEBREW_INSTALL_CLEANUP"]
+      return if Homebrew::EnvConfig.no_install_cleanup?
 
       cleanup = Cleanup.new
       if cleanup.periodic_clean_due?
         cleanup.periodic_clean!
-      elsif f.installed?
+      elsif f.latest_version_installed?
         cleanup.cleanup_formula(f)
       end
     end
 
     def periodic_clean_due?
-      return false if ENV["HOMEBREW_NO_INSTALL_CLEANUP"]
-      return unless ENV["HOMEBREW_INSTALL_CLEANUP"]
-      return true unless PERIODIC_CLEAN_FILE.exist?
+      return false if Homebrew::EnvConfig.no_install_cleanup?
+
+      unless PERIODIC_CLEAN_FILE.exist?
+        HOMEBREW_CACHE.mkpath
+        FileUtils.touch PERIODIC_CLEAN_FILE
+        return false
+      end
 
       PERIODIC_CLEAN_FILE.mtime < CLEANUP_DEFAULT_DAYS.days.ago
     end
@@ -180,25 +181,34 @@ module Homebrew
       return false unless periodic_clean_due?
 
       ohai "`brew cleanup` has not been run in #{CLEANUP_DEFAULT_DAYS} days, running now..."
-      clean!
+      clean!(quiet: true, periodic: true)
     end
 
-    def clean!
+    def clean!(quiet: false, periodic: false)
       if args.empty?
         Formula.installed.sort_by(&:name).each do |formula|
-          cleanup_formula(formula)
+          cleanup_formula(formula, quiet: quiet, ds_store: false, cache_db: false)
         end
         cleanup_cache
         cleanup_logs
-        cleanup_portable_ruby
         cleanup_lockfiles
-        return if dry_run?
-
-        cleanup_old_cache_db
-        rm_ds_store
         prune_prefix_symlinks_and_directories
-        HOMEBREW_CACHE.mkpath
-        FileUtils.touch PERIODIC_CLEAN_FILE
+
+        unless dry_run?
+          cleanup_cache_db
+          rm_ds_store
+          HOMEBREW_CACHE.mkpath
+          FileUtils.touch PERIODIC_CLEAN_FILE
+        end
+
+        # Cleaning up Ruby needs to be done last to avoid requiring additional
+        # files afterwards. Additionally, don't allow it on periodic cleans to
+        # avoid having to try to do a `brew install` when we've just deleted
+        # the running Ruby process...
+        return if periodic
+
+        cleanup_portable_ruby
+        cleanup_bootsnap
       else
         args.each do |arg|
           formula = begin
@@ -209,7 +219,7 @@ module Homebrew
 
           cask = begin
             Cask::CaskLoader.load(arg)
-          rescue Cask::CaskUnavailableError
+          rescue Cask::CaskError
             nil
           end
 
@@ -223,21 +233,23 @@ module Homebrew
       @unremovable_kegs ||= []
     end
 
-    def cleanup_formula(formula)
-      formula.eligible_kegs_for_cleanup.each(&method(:cleanup_keg))
+    def cleanup_formula(formula, quiet: false, ds_store: true, cache_db: true)
+      formula.eligible_kegs_for_cleanup(quiet: quiet)
+             .each(&method(:cleanup_keg))
       cleanup_cache(Pathname.glob(cache/"#{formula.name}--*"))
-      rm_ds_store([formula.rack])
+      rm_ds_store([formula.rack]) if ds_store
+      cleanup_cache_db(formula.rack) if cache_db
       cleanup_lockfiles(FormulaLock.new(formula.name).path)
     end
 
-    def cleanup_cask(cask)
+    def cleanup_cask(cask, ds_store: true)
       cleanup_cache(Pathname.glob(cache/"Cask/#{cask.token}--*"))
-      rm_ds_store([cask.caskroom_path])
+      rm_ds_store([cask.caskroom_path]) if ds_store
       cleanup_lockfiles(CaskLock.new(cask.token).path)
     end
 
     def cleanup_keg(keg)
-      cleanup_path(keg) { keg.uninstall }
+      cleanup_path(keg) { keg.uninstall(raise_failures: true) }
     rescue Errno::EACCES => e
       opoo e.message
       unremovable_kegs << keg
@@ -245,6 +257,7 @@ module Homebrew
 
     def cleanup_logs
       return unless HOMEBREW_LOGS.directory?
+
       logs_days = if days > CLEANUP_DEFAULT_DAYS
         CLEANUP_DEFAULT_DAYS
       else
@@ -277,6 +290,8 @@ module Homebrew
             # Skip incomplete downloads which are still in progress.
             next
           end
+        elsif download.directory?
+          FileUtils.rm_rf download
         else
           download.unlink
         end
@@ -287,6 +302,8 @@ module Homebrew
       entries ||= [cache, cache/"Cask"].select(&:directory?).flat_map(&:children)
 
       entries.each do |path|
+        next if path == PERIODIC_CLEAN_FILE
+
         FileUtils.chmod_R 0755, path if path.go_cache_directory? && !dry_run?
         next cleanup_path(path) { path.unlink } if path.incomplete?
         next cleanup_path(path) { FileUtils.rm_rf path } if path.nested_cache?
@@ -300,33 +317,33 @@ module Homebrew
           next
         end
 
-        next cleanup_path(path) { path.unlink } if path.stale?(scrub?)
+        # If we've specified --prune don't do the (expensive) .stale? check.
+        cleanup_path(path) { path.unlink } if !prune? && path.stale?(scrub: scrub?)
       end
 
       cleanup_unreferenced_downloads
     end
 
     def cleanup_path(path)
+      return unless path.exist?
       return unless @cleaned_up_paths.add?(path)
 
       disk_usage = path.disk_usage
 
       if dry_run?
         puts "Would remove: #{path} (#{path.abv})"
+        @disk_cleanup_size += disk_usage
       else
         puts "Removing: #{path}... (#{path.abv})"
         yield
+        @disk_cleanup_size += disk_usage - path.disk_usage
       end
-
-      @disk_cleanup_size += disk_usage
     end
 
     def cleanup_lockfiles(*lockfiles)
       return if dry_run?
 
-      if lockfiles.empty? && HOMEBREW_LOCKS.directory?
-        lockfiles = HOMEBREW_LOCKS.children.select(&:file?)
-      end
+      lockfiles = HOMEBREW_LOCKS.children.select(&:file?) if lockfiles.empty? && HOMEBREW_LOCKS.directory?
 
       lockfiles.each do |file|
         next unless file.readable?
@@ -341,61 +358,89 @@ module Homebrew
     end
 
     def cleanup_portable_ruby
-      system_ruby_version =
-        Utils.popen_read("/usr/bin/ruby", "-e", "puts RUBY_VERSION")
-             .chomp
-      use_system_ruby = (
-        Gem::Version.new(system_ruby_version) >= Gem::Version.new(RUBY_VERSION)
-      ) && ENV["HOMEBREW_FORCE_VENDOR_RUBY"].nil?
-      vendor_path = HOMEBREW_LIBRARY/"Homebrew/vendor"
-      portable_ruby_version_file = vendor_path/"portable-ruby-version"
-      portable_ruby_version = if portable_ruby_version_file.exist?
-        portable_ruby_version_file.read
-                                  .chomp
-      end
+      rubies = [which("ruby"), which("ruby", ENV["HOMEBREW_PATH"])].compact
+      system_ruby = Pathname.new("/usr/bin/ruby")
+      rubies << system_ruby if system_ruby.exist?
 
-      portable_ruby_path = vendor_path/"portable-ruby"
-      portable_ruby_glob = "#{portable_ruby_path}/*.*"
-      Pathname.glob(portable_ruby_glob).each do |path|
-        next if !use_system_ruby && portable_ruby_version == path.basename.to_s
-        if dry_run?
-          puts "Would remove: #{path} (#{path.abv})"
-        else
-          FileUtils.rm_rf path
+      use_system_ruby = if Homebrew::EnvConfig.force_vendor_ruby?
+        false
+      elsif OS.mac?
+        ENV["HOMEBREW_MACOS_SYSTEM_RUBY_NEW_ENOUGH"].present?
+      else
+        check_ruby_version = HOMEBREW_LIBRARY_PATH/"utils/ruby_check_version_script.rb"
+        rubies.uniq.any? do |ruby|
+          quiet_system ruby, "--enable-frozen-string-literal", "--disable=gems,did_you_mean,rubyopt",
+                       check_ruby_version, HOMEBREW_REQUIRED_RUBY_VERSION
         end
       end
 
-      return unless Dir.glob(portable_ruby_glob).empty?
-      return unless portable_ruby_path.exist?
+      vendor_dir = HOMEBREW_LIBRARY/"Homebrew/vendor"
+      portable_ruby_latest_version = (vendor_dir/"portable-ruby-version").read.chomp
 
-      bundler_path = vendor_path/"bundle/ruby"
+      portable_rubies_to_remove = []
+      Pathname.glob(vendor_dir/"portable-ruby/*.*").select(&:directory?).each do |path|
+        next if !use_system_ruby && portable_ruby_latest_version == path.basename.to_s
+
+        portable_rubies_to_remove << path
+        puts "Would remove: #{path} (#{path.abv})" if dry_run?
+      end
+
+      return if portable_rubies_to_remove.empty?
+
+      bundler_path = vendor_dir/"bundle/ruby"
       if dry_run?
-        puts "Would remove: #{bundler_path} (#{bundler_path.abv})"
-        puts "Would remove: #{portable_ruby_path} (#{portable_ruby_path.abv})"
+        puts Utils.popen_read("git", "-C", HOMEBREW_REPOSITORY, "clean", "-nx", bundler_path).chomp
       else
-        FileUtils.rm_rf [bundler_path, portable_ruby_path]
+        puts Utils.popen_read("git", "-C", HOMEBREW_REPOSITORY, "clean", "-ffqx", bundler_path).chomp
+      end
+
+      return if dry_run?
+
+      FileUtils.rm_rf portable_rubies_to_remove
+    end
+
+    def cleanup_bootsnap
+      bootsnap = cache/"bootsnap"
+      return unless bootsnap.exist?
+
+      if dry_run?
+        puts "Would remove: #{bootsnap} (#{bootsnap.abv})"
+      else
+        FileUtils.rm_rf bootsnap
       end
     end
 
-    def cleanup_old_cache_db
+    def cleanup_cache_db(rack = nil)
       FileUtils.rm_rf [
         cache/"desc_cache.json",
         cache/"linkage.db",
         cache/"linkage.db.db",
       ]
+
+      CacheStoreDatabase.use(:linkage) do |db|
+        break unless db.created?
+
+        db.each_key do |keg|
+          next if rack.present? && !keg.start_with?("#{rack}/")
+          next if File.directory?(keg)
+
+          LinkageCacheStore.new(keg, db).delete!
+        end
+      end
     end
 
     def rm_ds_store(dirs = nil)
-      dirs ||= begin
-        Keg::MUST_EXIST_DIRECTORIES + [
-          HOMEBREW_PREFIX/"Caskroom",
-        ]
-      end
-      dirs.select(&:directory?).each.parallel do |dir|
-        system_command "find",
-          args:         [dir, "-name", ".DS_Store", "-delete"],
-          print_stderr: false
-      end
+      dirs ||= Keg::MUST_EXIST_DIRECTORIES + [
+        HOMEBREW_PREFIX/"Caskroom",
+      ]
+      dirs.select(&:directory?)
+          .flat_map { |d| Pathname.glob("#{d}/**/.DS_Store") }
+          .each do |dir|
+            dir.unlink
+          rescue Errno::EACCES
+            # don't care if we can't delete a .DS_Store
+            nil
+          end
     end
 
     def prune_prefix_symlinks_and_directories
@@ -410,9 +455,7 @@ module Homebrew
           path.extend(ObserverPathnameExtension)
           if path.symlink?
             unless path.resolved_path_exists?
-              if path.to_s =~ Keg::INFOFILE_RX
-                path.uninstall_info unless dry_run?
-              end
+              path.uninstall_info if path.to_s.match?(Keg::INFOFILE_RX) && !dry_run?
 
               if dry_run?
                 puts "Would remove (broken link): #{path}"
@@ -420,7 +463,7 @@ module Homebrew
                 path.unlink
               end
             end
-          elsif path.directory? && !Keg::MUST_EXIST_SUBDIRECTORIES.include?(path)
+          elsif path.directory? && Keg::MUST_EXIST_SUBDIRECTORIES.exclude?(path)
             dirs << path
           end
         end
